requirements.txt
fastapi
dotenv
google
pillow
io
pydantic
uvicorn